query params for clothing items and auth.user for clothing items

+++ girlsclothesapi/views/clothing_item.py
@@ -10,31 +10,80 @@ class ClothingItemView(ViewSet):
 
     def list(self, request):
 
-        item = ClothingItem.objects.all()
-        serializer = ClothingItemSerializer(item, many=True)
+        items = []
+
+        if request.auth.user.is_staff:
+            items = ClothingItem.objects.all()
+
+            if "type" in request.query_params:
+                items = items.filter(clothing_type_id=request.query_params['type'])
+                if request.query_params['type'] == "{id}":
+                    pass
+
+            if "isClean" in request.query_params:
+                items = items.filter(clean_or_dirty=True)
+                if request.query_params['isClean'] == True:
+                    pass
+
+            if "siblingMatch" in request.query_params:
+                items = items.filter(sibling_has_match=True)
+                if request.query_params['siblingMatch'] == True:
+                    pass
+
+            if "kid" in request.query_params:
+                items = items.filter(kid_id=request.query_params['kid'])
+                if request.query_params['kid'] == "{id}":
+                    pass
+
+        else:
+            items = ClothingItem.objects.filter(kid__user=request.auth.user)
+
+            if "type" in request.query_params:
+                items = items.filter(clothing_type_id=request.query_params['type'])
+                if request.query_params['type'] == "{id}":
+                    pass
+
+            if "isClean" in request.query_params:
+                items = items.filter(clean_or_dirty=True)
+                if request.query_params['isClean'] == True:
+                    pass
+
+            if "siblingMatch" in request.query_params:
+                items = items.filter(sibling_has_match=True)
+                if request.query_params['siblingMatch'] == True:
+                    pass
+
+        
+        serializer = ClothingItemSerializer(items, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def retrieve(self, request, pk):
 
-        item = ClothingItem.objects.get(pk=pk)
+        if request.auth.user.is_staff:
+            item = ClothingItem.objects.get(pk=pk)
+
+        else:
+            item = ClothingItem.objects.filter(kid__user=request.auth.user)
+
         serializer = ClothingItemSerializer(item)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def create(self, request):
 
-        clothing_type= ClothingType.objects.get(pk=request.data["clothing_type"])
-        kid = Kid.objects.get(pk=request.data["kid"])
+        if request.auth.user.is_staff:
+            clothing_type= ClothingType.objects.get(pk=request.data["clothing_type"])
+            kid = Kid.objects.get(pk=request.data["kid"])
 
-        item = ClothingItem.objects.create(
-            item_description=request.data["item_description"],
-            size=request.data["size"],
-            clean_or_dirty=request.data["clean_or_dirty"],
-            item_fits=request.data["item_fits"],
-            sibling_has_match=request.data["sibling_has_match"],
-            item_image=request.data["item_image"],
-            clothing_type=clothing_type,
-            kid=kid
-        )
+            item = ClothingItem.objects.create(
+                item_description=request.data["item_description"],
+                size=request.data["size"],
+                clean_or_dirty=request.data["clean_or_dirty"],
+                item_fits=request.data["item_fits"],
+                sibling_has_match=request.data["sibling_has_match"],
+                item_image=request.data["item_image"],
+                clothing_type=clothing_type,
+                kid=kid
+            )
 
         serializer = ClothingItemSerializer(item)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -58,30 +107,34 @@ class ClothingItemView(ViewSet):
         return Response(None, status=status.HTTP_204_NO_CONTENT)
 
     def destroy(self, request, pk):
-        item = ClothingItem.objects.get(pk=pk)
-        item.delete()
+        
+        if request.auth.user.is_staff:
+            item = ClothingItem.objects.get(pk=pk)
+            item.delete()
         return Response(None, status=status.HTTP_204_NO_CONTENT)
 
     @action(methods=['post'], detail=True)
     def adduses(self, request, pk):
 
-        use = ClothingUse.objects.get(pk=request.data['clothing_uses'])
-        outfit = ClothingItem.objects.get(pk=pk)
-        outfit.clothing_uses.add(use)
+        if request.auth.user.is_staff:
+            use = ClothingUse.objects.get(pk=request.data['clothing_uses'])
+            outfit = ClothingItem.objects.get(pk=pk)
+            outfit.clothing_uses.add(use)
         return Response({'message': 'Clothing use added to this outfit'}, status=status.HTTP_201_CREATED)
 
     @action(methods=['delete'], detail=True)
     def removeuses(self, request, pk):
 
-        use = ClothingUse.objects.get(pk=request.data['clothing_uses'])
-        outfit = ClothingItem.objects.get(pk=pk)
-        outfit.clothing_uses.remove(use)
+        if request.auth.user.is_staff:
+            use = ClothingUse.objects.get(pk=request.data['clothing_uses'])
+            outfit = ClothingItem.objects.get(pk=pk)
+            outfit.clothing_uses.remove(use)
         return Response({'message': 'Clothing use removed from this outfit'}, status=status.HTTP_204_NO_CONTENT)
 
 class ClothingItemSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = ClothingItem
-        fields = ('id', 'item_description', 'clothing_type', 'kid', 'size', 'clean_or_dirty', 'item_fits', 'sibling_has_match', 'item_image', 'clothing_uses')
-        depth = 2
+        fields = ('id', 'item_description', 'clothing_type', 'kid_nickname', 'size', 'clean_or_dirty', 'item_fits', 'sibling_has_match', 'item_image', 'clothing_uses')
+        depth = 1
